version sans nombre de screens fixe, on passe à juste un nombre d'images affichées

+++ app/prefs.php
@@ -26,25 +26,13 @@ class PrefsManager {
 		}
 		$this->file = $file;
 		// init
-		if(isset($this->prefs['Nscreens'])) {
-			if(intval($this->prefs['Nscreens'])>0) {
-				$Nscreens = $this->prefs['Nscreens'];
-			}
-		}
 		if(isset($this->prefs['tweets']) && is_array($this->prefs['tweets'])) {
 			$this->tweets = $this->prefs['tweets'];
 		}
 		if(isset($this->prefs['screens']) && is_array($this->prefs['screens'])) {
 			$this->screens = $this->prefs['screens'];
 			// limiter le nombre de screens à $Nscreens
-			if($Nscreens>0) {
-				$this->screens = array_filter($this->screens,
-					function($n) use ($Nscreens) {
-						return $n>0 && $n<=$Nscreens;
-					},
-					ARRAY_FILTER_USE_KEY
-				);
-			}
+			$this->screens = array_slice($this->screens,0,max(1,$Nscreens));
 		}
 		if(isset($this->prefs['scores']) && is_array($this->prefs['scores'])) {
 			$this->scores = $this->prefs['scores'];
@@ -100,13 +88,30 @@ class PrefsManager {
 		}
 		return "";
 	}
+	function addScreen($file,$top,$left,$zoom,$stamp=0) {
+		$this->setScreen(null,$file,$top,$left,$zoom,$stamp);
+	}
+	function insertScreen($screenIns,$file,$top,$left,$zoom) {
+		$stamp = time();
+		$screen = array(
+			"file" => $file,
+			"top" => $top,
+			"left" => $left,
+			"zoom" => $zoom,
+			"stamp" => $stamp,
+		);
+		array_splice($this->screens,$screenIns,0,[$screen]);
+	}
 	function setScreen($screenNum,$file,$top,$left,$zoom,$stamp=0) {
 		if($stamp == 0) {
 			$stamp = time();
 		}
+		$oldfile = "";
 		if(isset($this->screens[$screenNum])) {
 			$screen = $this->screens[$screenNum];
+			$oldfile = $screen['file'];
 		} else {
+			$screenNum = count($this->screens);
 			$screen = array(
 				"file" => $file,
 				"top" => 0,
@@ -121,6 +126,21 @@ class PrefsManager {
 		if($zoom != null) $screen['zoom'] = $zoom;
 		if($stamp != null) $screen['stamp'] = $stamp;
 		$this->screens[$screenNum] = $screen;
+		# supprimer le fichier remplacé
+		if($oldfile != "") {
+			foreach($this->screens as $screen) {
+				if($screen['file'] == $oldfile) {
+					return;
+				}
+			}
+			if(file_exists(SCREENS_DIR.$oldfile)) {
+				unlink(SCREENS_DIR.$oldfile);
+			}
+		}
+	}
+	// nombre de screens
+	function screenCount() {
+		return count($this->screens);
 	}
 	function screenFile($screenNum) {
 		if(isset($this->screens[$screenNum])) {
@@ -172,25 +192,28 @@ class PrefsManager {
 		}
 		return 0;
 	}
+	function switch_screens($screen1,$screen2) {
+		if(isset($this->screens[$screen1]) && isset($this->screens[$screen2])) {
+			$theScreen = $this->screens[$screen1];
+			$this->screens[$screen1] = $this->screens[$screen2];
+			$this->screens[$screen2] = $theScreen;
+		}
+	}
 	function effacer_screen($screen) {
-		$file = $this->screenFile($screen);
-		if($file != "") {
-			$this->screens[$screen] = array(
-				'file' => "",
-				'top' => 0,
-				'left' => 0,
-				'zoom' => 0,
-				'stamp' => 0,
-			);
+		if(isset($this->screens[$screen])) {
+			$file = $this->screenFile($screen);
+			array_splice($this->screens,$screen,1);
 			$this->save();
 			// n'effacer que si l'image n'est pas dans un autre screen
-			foreach($this->screens as $screen) {
-				if($screen['file'] == $file) {
-					return;
+			if($file != "") {
+				foreach($this->screens as $screen) {
+					if($screen['file'] == $file) {
+						return;
+					}
 				}
-			}
-			if(file_exists(SCREENS_DIR.$file)) {
-				unlink(SCREENS_DIR.$file);
+				if(file_exists(SCREENS_DIR.$file)) {
+					unlink(SCREENS_DIR.$file);
+				}
 			}
 		}
 	}
